removed some comment.

+++ uk/ac/babraham/BamQC/Modules/IndelFrequencies.java
@@ -220,14 +220,14 @@ public class IndelFrequencies extends AbstractQCModule {
 
 	@Override	
 	public boolean raisesError() {
-		if(firstMaxY+secondMaxY > ModuleConfig.getParam("variant_call_position_indel_threshold", "error"))
+		if(firstMaxY+secondMaxY > ModuleConfig.getParam("variant_call_position_indel_threshold", "error").doubleValue())
 			return true;		
 		return false;
 	}
 
 	@Override	
 	public boolean raisesWarning() {
-		if(firstMaxY+secondMaxY > ModuleConfig.getParam("variant_call_position_indel_threshold", "warn"))
+		if(firstMaxY+secondMaxY > ModuleConfig.getParam("variant_call_position_indel_threshold", "warn").doubleValue())
 			return true;		
 		return false;
 	}
