+++ uk/ac/babraham/BamQC/Modules/SNPFrequencies.java
@@ -197,14 +197,14 @@ public class SNPFrequencies extends AbstractQCModule {
 
 	@Override	
 	public boolean raisesError() {
-		if(firstMaxY+secondMaxY > ModuleConfig.getParam("variant_call_position_snp_threshold", "error"))
+		if(firstMaxY+secondMaxY > ModuleConfig.getParam("variant_call_position_snp_threshold", "error").doubleValue())
 			return true;		
 		return false;
 	}
 
 	@Override	
 	public boolean raisesWarning() {
-		if(firstMaxY+secondMaxY > ModuleConfig.getParam("variant_call_position_snp_threshold", "warn"))
+		if(firstMaxY+secondMaxY > ModuleConfig.getParam("variant_call_position_snp_threshold", "warn").doubleValue())
 			return true;		
 		return false;
 	}
